Break out test code into module for Threads, Runs, Steps, Messages
